move proximity code, for reals fix failing check
